% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_indicators.R
\name{calculate_participation}
\alias{calculate_participation}
\title{Calculate key indicators}
\usage{
calculate_participation(
  x,
  disaggregated = FALSE,
  level = NA,
  years = NA,
  ros = "all",
  types = "Actual"
)
}
\arguments{
\item{x}{a data.frame in the format}

\item{disaggregated}{logical. If FALSE (default), return totals for all levels. If TRUE, return disaggregate totals at the specified levels (IM, OU, and RO).}
}
\value{
data.frame with combined ftfms and DIS data
}
\description{
create data set from DIS exports and FTFMS data
}
