% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_indicators.R
\name{combine_indicators}
\alias{combine_indicators}
\title{Combine legacy and current indicators}
\usage{
combine_indicators(x, ...)
}
\arguments{
\item{x}{A data.frame in the format of indicators_df. Required columns are ro, ou, code, and year.}
}
\description{
This function combines legacy indicators with the current FTF indicators.
The names of the indicators are also adjusted to reflect the combination. Adds
a new column named indicator_name.
}
