% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disR-package.R
\docType{package}
\name{disR-package}
\alias{disR}
\alias{disR-package}
\title{disR: What the Package Does (One Line, Title Case)}
\description{
This package facilitates summarizing and managing data extracts from the Development Information System, at the Implementing Mechanism level of disaggregation.
}
\author{
\strong{Maintainer}: Ryan Thomas \email{rythomas@usaid.gov} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
