% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_indicators.R
\name{filter_indicators}
\alias{filter_indicators}
\title{Filter indicators with keywords}
\usage{
filter_indicators(
  x,
  ic_1819 = NA,
  d1 = NA,
  d2 = NA,
  d3 = NA,
  d4 = NA,
  search = NA,
  ...
)
}
\description{
Filter indicators with keywords
}
