% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_filter_magnus.R
\name{filter_magnus}
\alias{filter_magnus}
\title{Filter indicators with keywords}
\usage{
filter_magnus(
  x,
  ic_1819 = NA,
  d1 = NA,
  d2 = NA,
  d3 = NA,
  d4 = NA,
  search = NA,
  ...
)
}
\description{
Filter indicators with keywords
}
