% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_udns.R
\name{filter_udns}
\alias{filter_udns}
\title{Functions to parse udns that are the unique udns to calculate indicator totals}
\usage{
filter_udns(ic, udn)
}
\description{
Functions to parse udns that are the unique udns to calculate indicator totals
}
