% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_indicators.R
\name{gender_financing_}
\alias{gender_financing_}
\title{Calculate all the priority target values}
\usage{
gender_financing_(x, level = "im")
}
\value{
a list of two data frames, one with the summary of priority targets and one with all the data for each indicator nested by activity
}
\description{
Takes an export from DIS OU Activity Indicator Results Report
}
