% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets4_formats.R
\name{get_cells}
\alias{get_cells}
\title{extensions for openxlsx
#' #export textStyle
textStyle <- createStyle(
halign = "left",
valign = "top",
wrapText = TRUE,
)}
\usage{
get_cells(data, target_string, column_indexes)
}
\description{
extensions for openxlsx
#' #export textStyle
textStyle <- createStyle(
halign = "left",
valign = "top",
wrapText = TRUE,
)
}
