% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load DIS and FTFMS data in one of two formats \"magnus\" or \"extract\".}
\usage{
load_data(database = "extract")
}
\description{
This function is a wrapper around sum(). The sum() function returns 0 for
vectors that have all NA values, which is not desired behavior especially
for the DIS analysis.
}
