% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_database.R
\name{make_database}
\alias{make_database}
\title{make ftf indicator database}
\usage{
make_database(
  input_dir = "../../indicators/basic",
  output_dir = "../data/",
  format = c("relational", "tidy"),
  version = Sys.Date()
)
}
\arguments{
\item{input_dir}{character. The directory with all of the indicator files in the same format (xlsx by default).}

\item{output_dir}{character. The directory where you want to store the output files.}

\item{format}{character. One or both (default) of c("relational", "tidy") to specify the format of the data. The default is "relational".}

\item{version}{character. The folder where to locate the output_dir. The default is \code{sys.Date()}.}

\item{input_pattern}{character. The file format for all input files.}
}
\value{
saves database in .rdata format in the specified directory.
}
\description{
create a single database from all the final indicator spreadsheets.
}
\examples{
make_database()

}
