% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets4_formats.R
\name{number_format}
\alias{number_format}
\title{Create numeric cell formats for googlesheets4 API}
\usage{
number_format(type, pattern)
}
\arguments{
\item{type}{one of "PERCENT", "CURRENCY", or "NUMBER"}

\item{pattern}{a regex pattern to apply to the number format. See API documentation: https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/cells#cellformat}
}
\description{
Create numeric cell formats for googlesheets4 API
}
\examples{
percent_cell_format  <- number_format(type = "PERCENT", pattern = "#0\%")
currency_cell_format <- number_format(type = "CURRENCY", sprintf('\%s#,##0', "$"))
number_cell_format   <- number_format(type = "NUMBER", '#,##0')
parity_cell_format   <- number_format(type = "CURRENCY", sprintf('\%s#,##0.00', "$"))
}
