% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_indicators.R
\name{priority_targets_}
\alias{priority_targets_}
\title{Calculate all the priority target values}
\usage{
priority_targets_(df, level = "im")
}
\value{
a list of two data frames, one with the summary of priority targets and one with all the data for each indicator nested by activity
}
\description{
Takes an export from DIS OU Activty Indicator Results Report
}
\examples{
priority_targets_(df)
}
