% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_dis_to_ftfms.R
\name{rbind_dis_to_ftfms}
\alias{rbind_dis_to_ftfms}
\title{parse DIS extracts into usable data}
\usage{
rbind_dis_to_ftfms(
  input_dir = "../../indicators/extract/",
  output_dir = "../data/"
)
}
\arguments{
\item{input_dir}{a directory with raw ftfms and dis (extract) data}

\item{output_dir}{a directory to store outputs}
}
\value{
data.frame with combined ftfms and DIS data
}
\description{
create data set from DIS exports and FTFMS data
}
\examples{
#' combine_extracts()
}
