% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_dis_to_ftfms.R
\name{read_dis}
\alias{read_dis}
\title{parse DIS extracts into usable data}
\usage{
read_dis(input)
}
\arguments{
\item{input_dir}{a directory with raw dis (extract CSV) data}

\item{output_dir}{a directory to store outputs}
}
\value{
data.frame with combined DIS data
}
\description{
create data set from DIS exports and FTFMS data
}
\examples{
#' dis <- read_dis()
}
