% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_indicators.R
\name{read_dis_extract}
\alias{read_dis_extract}
\title{parse DIS extracts into usable data}
\usage{
read_dis_extract(input, output_dir = NA)
}
\arguments{
\item{input}{character. a directory with raw dis (extract CSV) data}

\item{output_dir}{character. A directory to store outputs, or NA (default) to print output}
}
\value{
data.frame with combined DIS data
}
\description{
create data set from DIS exports and FTFMS data
}
\examples{
#' dis <- read_dis_extract()
}
