% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_indicators.R
\name{read_excel_dis}
\alias{read_excel_dis}
\title{Read DIS Excel reports from}
\usage{
read_excel_dis(input)
}
\arguments{
\item{input}{character. An excel file exported from DIS}
}
\value{
data.frame with DIS exported data
}
\description{
Read data set from DIS exports and rename the columns to be standardized
}
\examples{
#' dis <- read_excel_dis()
}
