% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_indicators.R
\name{read_ftfms}
\alias{read_ftfms}
\title{parse DIS extracts into usable data}
\usage{
read_ftfms(input = "../../indicators/extract/")
}
\arguments{
\item{input}{a directory with raw ftfms and dis (extract) data}
}
\value{
data.frame with combined ftfms data
}
\description{
create data set from DIS exports and FTFMS data
}
\examples{
#' ms <- read_ftfms()
}
