% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_indicator.R
\name{read_indicator}
\alias{read_indicator}
\title{Read indicator spreadsheets modified from the format of Magnus's Google Sheets}
\usage{
read_indicator(x, sheet = "IM Full Disaggs", version = "relational")
}
\description{
Read indicator spreadsheets modified from the format of Magnus's Google Sheets
}
