% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_id.R
\name{return_id}
\alias{return_id}
\title{Return unique IDs based on the row number.}
\usage{
return_id(row)
}
\arguments{
\item{used}{in mutate, a data.frame or vector}
}
\value{
Vector of IDs
}
\description{
This function is a helper used to write IDs for the FTF indicator database.
}
