% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_.R
\name{sum_}
\alias{sum_}
\title{Sum values in a vector and preserve NA if all are NA.}
\usage{
sum_(x)
}
\arguments{
\item{x}{A vector that you want to sum}
}
\description{
This function is a wrapper around sum(). The sum() function returns 0 for
vectors that have all NA values, which is not desired behavior especially
for the Feed the Future indicator analysis.
}
