% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_indicators.R
\name{sum_indicators}
\alias{sum_indicators}
\title{Summarize annual totals}
\usage{
sum_indicators(
  x,
  years = paste0("FY", 2011:2022),
  total = NA,
  disaggregate = NA,
  ind_code = NA,
  reporting_org = NA,
  ind_key = NA,
  ...
)
}
\description{
Summarize annual totals
}
