% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets4_formats.R
\name{text_format}
\alias{text_format}
\title{Create text cell formats for googlesheets4 API}
\usage{
text_format(strategy = "WRAP", valign = "TOP")
}
\arguments{
\item{strategy}{one of "WRAP" (default) or another wrapping strategy. See API documentation.}

\item{pattern}{a regex pattern to apply to the number format. See API documentation: https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/cells#cellformat}
}
\description{
Create text cell formats for googlesheets4 API
}
\examples{
wrap_cell_format     <- text_format(strategy = "WRAP", valign = "TOP")
}
