% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_disaggregates.R
\name{third_order}
\alias{third_order}
\title{Standardize FTFMS disaggregates}
\usage{
third_order(ic, ms_d1, ms_d2, ms_d3, ms_d4)
}
\arguments{
\item{ic}{a vector (column) of indicator codes}

\item{ms_d1}{a vector (column) of DisaggregationName1 from the FTFMS database}

\item{ms_d2}{a vector (column) of DisaggregationName2 from the FTFMS database}

\item{ms_d3}{a vector (column) of DisaggregationName3 from the FTFMS database}

\item{ms_d4}{a vector (column) of DisaggregationName4 from the FTFMS database}
}
\value{
a vector (column) of strings with disaggregates
}
\description{
These four functions are used in a mutate statement in 'rbind_dis_to_ftfms.R'. They use case_when to create standardized disaggregates for d1, d2, d3, d4 to match the DIS disaggregates.
}
\examples{
if(exists(ms)) ms \%>\% mutate(d3 = third_order())
}
