% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidyIndicators}
\alias{tidyIndicators}
\alias{indicators_df}
\alias{values}
\alias{ims}
\alias{indicators}
\alias{ftf_target_countries}
\title{Indicator database}
\format{
Five tibbles or data.frames, indicators_df, values, ims, indicators, ftf_target_countries

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 261278 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2578 rows and 5 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2837 rows and 8 columns.

An object of class \code{data.frame} with 20 rows and 2 columns.
}
\source{
\url{https://drive.google.com/drive/folders/1kNPQNmsWbbz_dRQ0KBHCzMFe-I4Djt78}
}
\usage{
indicators_df

values

ims

indicators

ftf_target_countries
}
\description{
Collated indicator database for Feed the Future indicators
}
\keyword{datasets}
