% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_indicators.R
\name{rbind_dis_to_ftfms}
\alias{rbind_dis_to_ftfms}
\title{parse DIS extracts into usable data}
\usage{
rbind_dis_to_ftfms(
  dis_input,
  ftfms_input = "../../indicators/extract/",
  output_dir = "../data/"
)
}
\arguments{
\item{dis_input}{character. File path of DIS extract CSV.}

\item{ftfms_input}{character. Directory path to ftfms Excel files}

\item{output_dir}{character. Directory path to store outputs combined_extracts.csv and combined_extracts.Rdata}
}
\value{
data.frame with combined ftfms and DIS data
}
\description{
create data set from DIS exports and FTFMS data
}
