% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_indicator.R
\name{read_indicator}
\alias{read_indicator}
\title{Read Google Sheets data and select columns}
\usage{
read_indicator(x, sheet = "IM Full Disaggs")
}
\description{
Read Google Sheets data and select columns
}
